+++ src/Wizard.js
@@ -35,6 +35,7 @@ export default class Wizard extends Webform {
     this.originalComponents = [];
     this.page = 0;
     this.currentPanel = null;
+    this.currentPanels = null;
     this.currentNextPage = 0;
     this._seenPages = [0];
     this.subWizards = [];
@@ -56,10 +57,15 @@ export default class Wizard extends Webform {
 
   getPages(args = {}) {
     const { all = false } = args;
-    const pages = this.pages
+    const pages = this.hasExtraPages ? this.components : this.pages;
+    const filteredPages = pages
       .filter(all ? _.identity : (p, index) => this._seenPages.includes(index));
 
-    return pages;
+    return filteredPages;
+  }
+
+  get hasExtraPages() {
+    return !_.isEmpty(this.subWizards) && !_.isEqual(this.pages, this.components);
   }
 
   get data() {
@@ -107,7 +113,6 @@ export default class Wizard extends Webform {
 
     this.on('subWizardsUpdated', () => {
       if (this.subWizards.length) {
-        this.transformPages();
         this.establishPages();
         this.redraw();
       }
@@ -168,7 +173,7 @@ export default class Wizard extends Webform {
     return {
       wizardKey: this.wizardKey,
       isBreadcrumbClickable: this.isBreadcrumbClickable(),
-      isSubForm: !!this.parent,
+      isSubForm: !!this.parent && !this.root?.component?.type === 'wizard',
       panels: this.allPages.length ? this.allPages.map(page => page.component) : this.pages.map(page => page.component),
       buttons: this.buttons,
       currentPage: this.page,
@@ -348,18 +353,37 @@ export default class Wizard extends Webform {
 
   transformPages() {
     const allComponents = [];
+    const currentPages = {};
+    const components = [];
     let defferedComponents = [];
     this.allPages = [];
+
+    if (this.components && this.components.length) {
+      this.components.map(page => {
+        if (page.component.type === 'panel') {
+          currentPages[page.component.key || page.component.title] = page;
+        }
+      });
+    }
+
+    this.originalComponents?.forEach((item) => {
+      if (!item.key) {
+        item.key = item.title;
+      }
+      if (currentPages[item.key]) {
+        components.push(currentPages[item.key]);
+      }
+    });
 
     // Get all components including all nested components and line up in the correct order
     const getAllComponents = (nestedComp, compsArr, pushAllowed = true) => {
       let hasNested = false;
       const nestedPages = [];
-      const components = nestedComp?.subForm ? nestedComp?.subForm.components : nestedComp.components || [];
-      const additionalComponents = components.filter(comp => !comp.subForm);
+      const components = nestedComp?.subForm ? nestedComp?.subForm.components : nestedComp?.components || [];
+      const additionalComponents = components.filter(comp => !comp.subForm && comp._visible);
 
       eachComponent(components, (comp) => {
-        if (comp.component.type === 'panel' && !getAllComponents(comp, compsArr, false)) {
+        if (comp.component.type === 'panel' && comp?.parent.wizard && !getAllComponents(comp, compsArr, false)) {
           if (pushAllowed) {
             nestedPages.push(comp);
           }
@@ -388,7 +412,7 @@ export default class Wizard extends Webform {
       return hasNested;
     };
 
-    this.components.forEach((component) => {
+    components.forEach((component) => {
       getAllComponents(component, allComponents);
     }, []);
 
@@ -446,6 +470,7 @@ export default class Wizard extends Webform {
       });
     }
 
+    this.transformPages();
     if (this.allPages && this.allPages.length) {
       this.pages = this.allPages;
     }
@@ -536,8 +561,7 @@ export default class Wizard extends Webform {
   }
 
   beforeSubmit() {
-    const hasExtraPages = !_.isEmpty(this.subWizards) && !_.isEqual(this.pages, this.components);
-    const pages = hasExtraPages ? this.getComponents() : this.getPages();
+    const pages = this.getPages();
 
     return NativePromise.all(pages.map((page) => {
       page.options.beforeSubmit = true;
@@ -658,6 +682,11 @@ export default class Wizard extends Webform {
 
   setValue(submission, flags = {}) {
     this._submission = submission;
+
+    if (flags && flags.fromSubmission && submission) {
+      this._data = submission.data;
+    }
+
     this.establishPages(submission.data);
     const changed = this.getPages({ all: true }).reduce((changed, page) => {
       return this.setNestedValue(page, submission.data, flags, changed) || changed;
@@ -722,10 +751,21 @@ export default class Wizard extends Webform {
     }
 
     // If the pages change, need to redraw the header.
-    const currentPanels = this.pages.map(page => page.component.key);
-    const panels = this.establishPages().map(panel => panel.key);
+    let currentPanels;
+    let panels;
     const currentNextPage = this.currentNextPage;
-    if (!_.isEqual(panels, currentPanels)) {
+    if (this.hasExtraPages) {
+      currentPanels = this.pages.map(page => page.component.key);
+      this.establishPages();
+      panels = this.pages.map(page => page.component.key);
+    }
+    else {
+      currentPanels = this.currentPanels || this.pages.map(page => page.component.key);
+      panels = this.establishPages().map(panel => panel.key);
+      this.currentPanels = panels;
+    }
+
+    if (!_.isEqual(panels, currentPanels) || (flags && flags.fromSubmission)) {
       this.redrawHeader();
     }
 
@@ -733,6 +773,13 @@ export default class Wizard extends Webform {
     if (currentNextPage !== this.getNextPage()) {
       this.redrawNavigation();
     }
+  }
+
+  redraw() {
+    if (this.parent?.component?.modalEdit) {
+      return this.parent.redraw();
+    }
+    return super.redraw();
   }
 
   checkValidity(data, dirty, row, currentPageOnly) {