+++ src/components/checkbox/Checkbox.js
@@ -19,7 +19,7 @@ export default class CheckBoxComponent extends Field {
       title: 'Checkbox',
       group: 'basic',
       icon: 'check-square',
-      documentation: 'http://help.form.io/userguide/#checkbox',
+      documentation: '/userguide/#checkbox',
       weight: 50,
       schema: CheckBoxComponent.schema()
     };